Move component classes to messages.py
