+++ hikari/internal/cache.py
@@ -74,7 +74,6 @@ if typing.TYPE_CHECKING:
     from hikari import traits
     from hikari import users as users_
     from hikari.interactions import base_interactions
-    from hikari.interactions import components as component_interactions
 
 ChannelT = typing.TypeVar("ChannelT", bound="channels_.GuildChannel")
 DataT = typing.TypeVar("DataT", bound="BaseData[typing.Any]")
@@ -773,7 +772,7 @@ class MessageData(BaseData[messages.Message]):
     referenced_message: typing.Optional[RefCell[MessageData]] = attr.field()
     interaction: typing.Optional[MessageInteractionData] = attr.field()
     application_id: typing.Optional[snowflakes.Snowflake] = attr.field()
-    components: typing.Tuple[component_interactions.PartialComponent, ...] = attr.field()
+    components: typing.Tuple[messages.PartialComponent, ...] = attr.field()
 
     @classmethod
     def build_from_entity(
@@ -888,6 +887,9 @@ class MessageData(BaseData[messages.Message]):
 
         if message.embeds is not undefined.UNDEFINED:
             self.embeds = tuple(map(_copy_embed, message.embeds))
+
+        if message.components is not undefined.UNDEFINED:
+            self.components = tuple(message.components)
 
         self.mentions.update(message.mentions, users=mention_users)
