+++ hikari/messages.py
@@ -36,6 +36,11 @@ __all__: typing.List[str] = [
     "MessageReference",
     "PartialMessage",
     "Message",
+    "ActionRowComponent",
+    "ButtonComponent",
+    "ButtonStyle",
+    "ComponentType",
+    "PartialComponent",
 ]
 
 import typing
@@ -62,7 +67,6 @@ if typing.TYPE_CHECKING:
     from hikari import users as users_
     from hikari.api import special_endpoints
     from hikari.interactions import base_interactions
-    from hikari.interactions import components as component_interactions
 
 _T = typing.TypeVar("_T")
 
@@ -475,6 +479,133 @@ class MessageInteraction:
     """Object of the user who invoked this interaction."""
 
 
+@typing.final
+class ComponentType(int, enums.Enum):
+    """Types of components found within Discord."""
+
+    ACTION_ROW = 1
+    """A non-interactive container component for other types of components.
+
+    !!! note
+        As this is a container component it can never be contained within another
+        component and therefore will always be top-level.l
+    """
+
+    BUTTON = 2
+    """A button component.
+
+    !!! note
+        This cannot be top-level and must be within a container component such
+        as `ComponentType.ACTION_ROW`.
+    """
+
+
+@typing.final
+class ButtonStyle(int, enums.Enum):
+    """Enum of the available button styles.
+
+    More information, such as how these look, can be found at
+    https://discord.com/developers/docs/interactions/message-components#buttons-button-styles
+    """
+
+    PRIMARY = 1
+    """A blurple "call to action" button."""
+
+    SECONDARY = 2
+    """A grey neutral button."""
+
+    SUCCESS = 3
+    """A green button."""
+
+    DANGER = 4
+    """A red button (usually indicates a destructive action)."""
+
+    LINK = 5
+    """A grey button which navigates to a URL.
+
+    !!! warning
+        Unlike the other button styles, clicking this one will not trigger an
+        interaction and custom_id shouldn't be included for this style.
+    """
+
+
+@attr.define(kw_only=True, weakref_slot=False)
+class PartialComponent:
+    """Base class for all component entities."""
+
+    type: typing.Union[ComponentType, int] = attr.field()
+    """The type of component this is."""
+
+
+@attr.define(hash=True, kw_only=True, weakref_slot=False)
+class ButtonComponent(PartialComponent):
+    """Represents a message button component.
+
+    !!! note
+        This is an embedded component and will only ever be found within
+        top-level container components such as `ActionRowComponent`.
+    """
+
+    style: typing.Union[ButtonStyle, int] = attr.field(eq=False)
+    """The button's style."""
+
+    label: typing.Optional[str] = attr.field(eq=False)
+    """Text label which appears on the button."""
+
+    emoji: typing.Optional[emojis_.Emoji] = attr.field(eq=False)
+    """Custom or unicode emoji which appears on the button."""
+
+    custom_id: typing.Optional[str] = attr.field(hash=True)
+    """Developer defined identifier for this button (will be >= 100 characters).
+
+    !!! note
+        This is required for the following button styles:
+
+        * `ButtonStyle.PRIMARY`
+        * `ButtonStyle.SECONDARY`
+        * `ButtonStyle.SUCCESS`
+        * `ButtonStyle.DANGER`
+    """
+
+    url: typing.Optional[str] = attr.field(eq=False)
+    """Url for `ButtonStyle.LINK` style buttons."""
+
+    is_disabled: bool = attr.field(eq=False)
+    """Whether the button is disabled."""
+
+
+@attr.define(weakref_slot=False)
+class ActionRowComponent(PartialComponent):
+    """Represents a row of components attached to a message.
+
+    !!! note
+        This is a top-level container component and will never be found within
+        another component.
+    """
+
+    components: typing.Sequence[PartialComponent] = attr.field()
+    """Sequence of the components contained within this row."""
+
+    @typing.overload
+    def __getitem__(self, index: int, /) -> PartialComponent:
+        ...
+
+    @typing.overload
+    def __getitem__(self, slice_: slice, /) -> typing.Sequence[PartialComponent]:
+        ...
+
+    def __getitem__(
+        self, index_or_slice: typing.Union[int, slice], /
+    ) -> typing.Union[PartialComponent, typing.Sequence[PartialComponent]]:
+        return self.components[index_or_slice]
+
+    def __iter__(self) -> typing.Iterator[PartialComponent]:
+        return iter(self.components)
+
+    def __len__(self) -> int:
+        return len(self.components)
+
+
 @attr_extensions.with_copy
 @attr.define(kw_only=True, repr=True, eq=False, weakref_slot=False)
 class PartialMessage(snowflakes.Unique):
@@ -618,9 +749,7 @@ class PartialMessage(snowflakes.Unique):
         This will only be provided for interaction messages.
     """
 
-    components: undefined.UndefinedOr[typing.Sequence[component_interactions.PartialComponent]] = attr.field(
-        hash=False, repr=False
-    )
+    components: undefined.UndefinedOr[typing.Sequence[PartialComponent]] = attr.field(hash=False, repr=False)
     """Sequence of the components attached to this message."""
 
     @property  # TODO: update this while refactoring message structure
@@ -1395,5 +1524,5 @@ class Message(PartialMessage):
         This will only be provided for interaction messages.
     """
 
-    components: typing.Sequence[component_interactions.PartialComponent]
+    components: typing.Sequence[PartialComponent]
     """Sequence of the components attached to this message."""
