+++ tests/hikari/interactions/test_components.py
@@ -32,33 +32,6 @@ def mock_app():
     return mock.Mock()
 
 
-class TestActionRowComponent:
-    def test_getitem_operator_with_index(self):
-        mock_component = object()
-        row = components.ActionRowComponent(type=1, components=[object(), mock_component, object()])
-
-        assert row[1] is mock_component
-
-    def test_getitem_operator_with_slice(self):
-        mock_component_1 = object()
-        mock_component_2 = object()
-        row = components.ActionRowComponent(type=1, components=[object(), mock_component_1, object(), mock_component_2])
-
-        assert row[1:4:2] == [mock_component_1, mock_component_2]
-
-    def test_iter_operator(self):
-        mock_component_1 = object()
-        mock_component_2 = object()
-        row = components.ActionRowComponent(type=1, components=[mock_component_1, mock_component_2])
-
-        assert list(row) == [mock_component_1, mock_component_2]
-
-    def test_len_operator(self):
-        row = components.ActionRowComponent(type=1, components=[object(), object()])
-
-        assert len(row) == 2
-
-
 class TestComponentInteraction:
     @pytest.fixture()
     def mock_component_interaction(self, mock_app):
